refactor(document): optimize delete flow with datagrid context menu and remove redundant dialog
