--- Prn212.AIStudyHub.WPF/Views/Documents/DeleteDocumentWindow.xaml
@@ -1,32 +1,0 @@
-<Window x:Class="Prn212.AIStudyHub.WPF.Views.Documents.DeleteDocumentWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        Title="Xóa Tài Liệu" Height="280" Width="400"
-        WindowStartupLocation="CenterOwner" ResizeMode="NoResize"
-        Background="{StaticResource BackgroundBrush}">
-    <Grid Margin="20">
-        <Grid.RowDefinitions>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="*"/>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="Auto"/>
-        </Grid.RowDefinitions>
-
-        <TextBlock Grid.Row="0" Text="Xóa Tài Liệu Học Tập" 
-                   Style="{StaticResource SubHeaderTextBlockStyle}" Margin="0,0,0,20"/>
-
-        <StackPanel Grid.Row="1" VerticalAlignment="Top">
-            <TextBlock Text="Chọn tài liệu muốn xóa (*):" Style="{StaticResource FormLabelStyle}"/>
-            <ComboBox x:Name="cbDocuments" Height="34" Margin="0,0,0,15"
-                      DisplayMemberPath="Title" SelectedValuePath="Id"
-                      Style="{StaticResource ModernComboBoxStyle}"/>
-        </StackPanel>
-
-        <UniformGrid Grid.Row="3" Columns="2" Rows="1" Height="38">
-            <Button x:Name="btnDelete" Content="Xóa Tài Liệu" Margin="0,0,10,0" Click="BtnDelete_Click"
-                    Style="{StaticResource DangerButtonStyle}"/>
-            <Button x:Name="btnCancel" Content="Đóng" Margin="10,0,0,0" Click="BtnCancel_Click"
-                    Style="{StaticResource ModernButtonStyle}"/>
-        </UniformGrid>
-    </Grid>
-</Window>
